sprint---11

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -6,12 +6,14 @@ import java.time.LocalDate;
 import java.util.HashSet;
 import java.util.Set;
 
+
 @Data
 public class Film {
-    private Integer id;
+    private long id;
     private String name;
     private String description;
     private LocalDate releaseDate;
-    private Integer duration;
-    private Set<Integer> likes = new HashSet<>();
+    private int duration;
+    private Set<Long> likes = new HashSet<>();
 }
+
