+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -6,12 +6,13 @@ import java.time.LocalDate;
 import java.util.HashSet;
 import java.util.Set;
 
+
 @Data
 public class User {
-    private Integer id;
+    private long id;
     private String email;
     private String login;
     private String name;
     private LocalDate birthday;
-    private Set<Integer> friends = new HashSet<>();
+    private Set<Long> friends = new HashSet<>();
 }
